Move new controlnet module to its own file
